+++ js/util.js
@@ -1,0 +1,27 @@
+// Генерируем рандомное целое число
+const getRandomInt = (min, max) => {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+// Генерируем рандомный ID
+function getRandomID(min, max) {
+  const previousValues = [];
+
+  return function () {
+
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    let currentValue = getRandomInt(min, max);
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInt(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+
+}
+
+export {getRandomID, getRandomInt};
